pyproject.toml
[tool.poetry]
authors = ["FL03 <jo3mccain@icloud.com>"]
description = "Synopsis is a complete file management application equipped with a number of unique extraction tools"
license = "MIT"
name = "synopsis"
version = "0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
Jinja2 = "latest"
arweave-python-client = "latest"
bcrypt = { extras = ["cryptography"], version = "latest" }
deta = "latest"
fastapi = "latest"
matplotlib = "latest"
numpy = "latest"
pandas = "latest"
panel = "latest"
plotly = "latest"
pydantic = "latest"
python = ">=3.9,<3.11"
python-dotenv = "latest"
python-jose = "latest"
strawberry-graphql = { extras = ["fastapi"], version = "latest"}
tortoise-orm = { extras = ["asyncpg"], version = "latest" }
uvicorn = "0.17.5"
web3 = "latest"

[tool.poetry.dev-dependencies]
black = "latest"
pytest = "latest"